Can the Clients list be simpler like the Sessions tab is. see attached.
So have the Owner Name as the bold part and then the dogs name underneath with a paw icon next to it.

Id like to add an icon next to the name too for if they are a Member which I would like to add as a checkbox field within the Client. When selected, this icon shows (see attached).