I need to add Address to the Clients details that will fill in if the Address section has been completed from the Behaviour Questionnaire. If it has filled in, then PostCode can be hidden as this is part of the Address.
I need to add an Active checkbox to Clients too. This will be what the Active counter tallies up with on the Dashboard so that needs to match up with the backend.

Also for the Upcoming Sessions on Dashboard this should only count those sessions coming up in the following week.