Any time the Owner Name and the Dog Name are mentioned together I'd like it displayed as a w/. So for example, Matthew McCartney w/ Kuki.
If the Client doesnt have a dog name added, then only show the Client Name.

Remove the headline cards at the top of the Dashboard - not needed.

Instead, a single narrow card that has details for the next session. Showing Owner and dog name, time and date of session, Next Session header and the Session Type.

Also, on the main calendar we now have on Dashboard, don't have the red background on the current day - but maybe a red ring around the border of the date.